goinventory file added
The initial goinventory.go file had  renamed to goinventory to reflect the package name. Services directory is too vauge and should follow the go convention of the package having the same name as the directory. This file stores the structs for our database tables as well as defining the interfaces for each one. Go mod file has also been updated to add in the new packages to reference.