Updated the migration file to alter the data attributes.
Count has been updated to not null has been added because that the count can be nullable. Same is true for the created_ at field. This value will be created on insert so should not be null.

The count default 0 has been removed because golang the defined empty value for type int is 0 already so we do not need this attribute on the field.