Updating the build workflow to manually get the go-migrate cli and also include ports for postgres service
The docker run didn't work I suspect because I didn't have the cli proper so now I get the installation as a step to run the migration command. Since this is only build, I do not need to save the credentials as a secret. Github will handle the vm which shouldn't be accessible when it finishes.