Fix: Deploy workflow did not recognize the sslmode query (#3)
* Committing go modules initialization

Initial commit for go mod file. The go mod file will list dependencies required for all packages.

* Created inventory interfaces and models

Created the inventory database model structs along with the interface methods.

* Initial migration files

Initial migration to postgres db for the inventory list.

* goinventory file added

The initial goinventory.go file had  renamed to goinventory to reflect the package name. Services directory is too vauge and should follow the go convention of the package having the same name as the directory. This file stores the structs for our database tables as well as defining the interfaces for each one. Go mod file has also been updated to add in the new packages to reference.

* Interface implementations of database structs.

Use the sqlx package to make the database calls for the interface implementations. Store.go is the file that pulls together the structs to be used later on.

* Updated the migration file to alter the data attributes.

Count has been updated to not null has been added because that the count can be nullable. Same is true for the created_ at field. This value will be created on insert so should not be null.

The count default 0 has been removed because golang the defined empty value for type int is 0 already so we do not need this attribute on the field.

* Remove inventory_list table first for the migration down script.

I had to change the order of the down due to the foreign key constraint on the inventory_item table first which would prevent it from being dropped first.

* Updating build workflow to work for feature/ branch naming convention.

The build workflow was not launching when I had been making commits on this branch. Reason for this is because the * signifier is only for branch names that do not include a slash.

* WIP: Adding postgres DB to build workflow to test run migrations

Trying to add a postgres service to action workflow to test run migrations.

* Adding foundation branch to build action.

I will be developing new applications features for the feature branch but I am currently using the foundation branch to get the core requirements up and running.

* Updating the build workflow to manually get the go-migrate cli and also include ports for postgres service

The docker run didn't work I suspect because I didn't have the cli proper so now I get the installation as a step to run the migration command. Since this is only build, I do not need to save the credentials as a secret. Github will handle the vm which shouldn't be accessible when it finishes.

* Removing a dash which is concatenating the command

* Adding sudo to get go migrate step.

* Updating build workflow to a simpler step.

* Debug - splitting out the get go-migrate cli into two steps.

This is to better understand where the error is coming from for /usr/bin/sudo: /usr/bin/sudo: cannot execute binary file
Error: Process completed with exit code 126.

* Updating build workflow to use the current branch that is has pushed.

* Use checkout actions v3

* Update the directory path to migrations directory.

* Isolating handler function into separate package.

Visual studio code had issues where the handler function was throwing errors if I had the function along with other packages. The docs didn't help too much because they just referenced sub packages whereas I wanted to use separate packages. The solution is to make the handler function into its own function package to match the package name used as part of golang convention.

* Deploy workflow db migration step added.

Database migration step added to make migrations to remote database. Secret values added to github action as well.

* Fix: use ghcr for image to allow the workflow to pass when publishing

* Fix: deploy workflow did not recognize the sslmode query

This should be required when making a migration to remote database.