Update first migration to use unix timestamps rather than date time.
Converting the datatype for the datetimes to big int to store unix timestamp. This is because a timestamp would be more flexible than a date time string through out the codebase due to parsing the values to specific formats from go and postgres and then in vuejs when I eventually get to it.

The api should just hand over the time stamp and the eventual front end should handle the view for that to the user.