Update inventory list added
For this update. This changes the updated at type to not null. The struct values for updated and created at also have json tags so they can be encoded.

Migrations have been updated to have the updated at field to be not null. Will default to 0.

WIP:
I decided to add a response struct in the handler to only allow for certain values to be returned in json. I figure that I dont need to return the entire json model back. The new piece of data will be the updated at value. The front end will need to check for the http response and get this new value.