Handlers - Inventory list (#13)
This pull request will migrate the faas-cli template to dockerfile to serve a chi server. Therefore, the start point of my application is main.go.

Secret definition added for the postgres database. This has been applied to the remote pi3 instance. I may want to split this out into host, port and user name. I made the secret just for the entire conn string for simplicity.

Unix timestamp column is now used for date columns. This required the migration to be changed by dropping the timestamp columns and recreating them. 

CRUD handlers for inventory list created. JSON validation package was added to validate the parsed json request.

As part of this pull, I also updated the readme to include some additional information about this project.