Infrastructure/auto migration (#14)
* Auto migration on start up.

Auto migration now happens when creating the NewStore func. I was debating putting this in main but ultimately decided not to since the migration is happening on for postgres specifically I guess.

* Remove build steps for getting go migrate cli

steps to get and execute migration command manually removed because the migration is now done in app on stat up.

* Updating the branches for build action to include current branch "infrastructure"

* Splitting out the action workflows into staging and prod

The repo secrets had to be altered for both release types.