This should make the migration folder accessible for the auto migrate (#18)
* Auto migration on start up.

Auto migration now happens when creating the NewStore func. I was debating putting this in main but ultimately decided not to since the migration is happening on for postgres specifically I guess.

* Remove build steps for getting go migrate cli

steps to get and execute migration command manually removed because the migration is now done in app on stat up.

* Updating the branches for build action to include current branch "infrastructure"

* Splitting out the action workflows into staging and prod

The repo secrets had to be altered for both release types.

* Test full pipeline deployment with test migration

* Fix: update the secret values used for workflows

* Updating the secret names in prod workflow to match the prod secrets

* Move migration folder to docker image /home/app

This will allow me to reference the migration folder in openfaas container.