Deploy prod auto migrations (#24)
* Fix: workflow secret names (#15) (#16)

Secret names for release fixed for workflows. This request also includes a test migration on a test table to ensure that deployment is okay.

* Release auto migration fix to staging environment (#19)

Currently, the service does not start up when auto migrating because faasd cannot find the migrations folder. Deployment should be tested that everything is okay by checking the dashboard for the service to be ready.

* Revert "Release auto migration fix to staging environment (#19)" (#20)

This reverts commit f95fb7fa9c5ee46b262ad9e8cfd917ce2d0d0501.

* Deploy: auto migration fix (#23)

* Fix: workflow secret names (#15)

Secret names for release fixed for workflows. This request also includes a test migration on a test table to ensure that deployment is okay.

* This should make the migration folder accessible for the auto migrate (#18)

* Auto migration on start up.

Auto migration now happens when creating the NewStore func. I was debating putting this in main but ultimately decided not to since the migration is happening on for postgres specifically I guess.

* Remove build steps for getting go migrate cli

steps to get and execute migration command manually removed because the migration is now done in app on stat up.

* Updating the branches for build action to include current branch "infrastructure"

* Splitting out the action workflows into staging and prod

The repo secrets had to be altered for both release types.

* Test full pipeline deployment with test migration

* Fix: update the secret values used for workflows

* Updating the secret names in prod workflow to match the prod secrets

* Move migration folder to docker image /home/app

This will allow me to reference the migration folder in openfaas container.

* Fix: Auto migration (#22)

* Auto migration on start up.

Auto migration now happens when creating the NewStore func. I was debating putting this in main but ultimately decided not to since the migration is happening on for postgres specifically I guess.

* Remove build steps for getting go migrate cli

steps to get and execute migration command manually removed because the migration is now done in app on stat up.

* Updating the branches for build action to include current branch "infrastructure"

* Splitting out the action workflows into staging and prod

The repo secrets had to be altered for both release types.

* Test full pipeline deployment with test migration

* Fix: update the secret values used for workflows

* Updating the secret names in prod workflow to match the prod secrets

* Move migration folder to docker image /home/app

This will allow me to reference the migration folder in openfaas container.

* Update to multi arch build added. Migration fixed to create test table

Dockerfile required additional changes to build for multi arch for arm.

The migration file had an error with a comma being on the end of the create table line. I also added a if not exists check to ensure that the migration is re runnable if it fails.